pyproject.toml
[project]
name = "sipin-meemoo-sip-2-validator"
version = "0.0.1"
description = "A service that validates incoming meemoo SIPs."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "viaa-chassis",
    "pulsar-client",
    "meemoo-cloudevents",
    "meemoo-sip-validator"
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "pytest-cov",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--rootdir=./app/ -v"
testpaths = ["tests"]
